Moving Machine to vm package
